refact(cmd): Move SetupControllerPublicAddress to config.Config
This moves this functionality from a method on a base.Server that
would take a config, to a method on a config.Config. This method
operated completely on the passed in config.Config, and did not use the
base.Server so there was no need for it to be in the base package.
