+++ internal/cmd/base/server_test.go
@@ -286,316 +286,6 @@ func TestServer_AddEventerToContext(t *testing.T) {
 	}
 }
 
-func TestSetupControllerPublicClusterAddress(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name                    string
-		inputConfig             *config.Config
-		inputFlagValue          string
-		stateFn                 func(t *testing.T)
-		expErr                  bool
-		expErrStr               string
-		expPublicClusterAddress string
-	}{
-		{
-			name: "nil controller",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: nil,
-			},
-			inputFlagValue:          "",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: ":9201",
-		},
-		{
-			name: "setting public cluster address directly with ip",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: "127.0.0.1",
-				},
-			},
-			inputFlagValue:          "",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "setting public cluster address directly with ip:port",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: "127.0.0.1:8080",
-				},
-			},
-			inputFlagValue:          "",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:8080",
-		},
-		{
-			name: "setting public cluster address to env var",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: "env://TEST_ENV_VAR_FOR_CONTROLLER_ADDR",
-				},
-			},
-			inputFlagValue: "",
-			stateFn: func(t *testing.T) {
-				t.Setenv("TEST_ENV_VAR_FOR_CONTROLLER_ADDR", "127.0.0.1:8080")
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:8080",
-		},
-		{
-			name: "setting public cluster address to env var that points to template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: "env://TEST_ENV_VAR_FOR_CONTROLLER_ADDR",
-				},
-			},
-			inputFlagValue: "",
-			stateFn: func(t *testing.T) {
-				t.Setenv("TEST_ENV_VAR_FOR_CONTROLLER_ADDR", `{{ GetAllInterfaces | include "flags" "loopback" | include "type" "IPV4" | join "address" " " }}`)
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "setting public cluster address to ip template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: `{{ GetAllInterfaces | include "flags" "loopback" | include "type" "IPV4" | join "address" " " }}`,
-				},
-			},
-			inputFlagValue:          "",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "setting public cluster address to multiline ip template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{
-					PublicClusterAddr: `{{ with $local := GetAllInterfaces | include "flags" "loopback" | include "type" "IPV4" -}}
-					{{- $local | join "address" " " -}}
-				  {{- end }}`,
-				},
-			},
-			inputFlagValue:          "",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "using flag value with ip only",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          "127.0.0.1",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "using flag value with ip:port",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          "127.0.0.1:8080",
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:8080",
-		},
-		{
-			name: "using flag value with ip template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          `{{ GetAllInterfaces | include "flags" "loopback" | include "type" "IPV4" | join "address" " " }}`,
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "using flag value with multiline ip template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue: `{{ with $local := GetAllInterfaces | include "flags" "loopback" | include "type" "IPV4" -}}
-			  {{- $local | join "address" " " -}}
-			{{- end }}`,
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "using flag value to point to env var with ip only",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue: "env://TEST_ENV_VAR_FOR_CONTROLLER_ADDR",
-			stateFn: func(t *testing.T) {
-				t.Setenv("TEST_ENV_VAR_FOR_CONTROLLER_ADDR", "127.0.0.1")
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "using flag value to point to env var with ip:port",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue: "env://TEST_ENV_VAR_FOR_CONTROLLER_ADDR",
-			stateFn: func(t *testing.T) {
-				t.Setenv("TEST_ENV_VAR_FOR_CONTROLLER_ADDR", "127.0.0.1:8080")
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:8080",
-		},
-		{
-			name: "read address from listeners ip only",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{
-						{Purpose: []string{"cluster"}, Address: "127.0.0.1"},
-					},
-				},
-				Controller: &config.Controller{},
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:9201",
-		},
-		{
-			name: "read address from listeners ip:port",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{
-						{Purpose: []string{"cluster"}, Address: "127.0.0.1:8080"},
-					},
-				},
-				Controller: &config.Controller{},
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: "127.0.0.1:8080",
-		},
-		{
-			name: "read address from listeners is ignored on different purpose",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{
-						{Purpose: []string{"somethingelse"}, Address: "127.0.0.1:8080"},
-					},
-				},
-				Controller: &config.Controller{},
-			},
-			expErr:                  false,
-			expErrStr:               "",
-			expPublicClusterAddress: ":9201",
-		},
-		{
-			name: "using flag value to point to nonexistent file",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          "file://this_doesnt_exist_for_sure",
-			expErr:                  true,
-			expErrStr:               "Error parsing public cluster addr: error reading file at file://this_doesnt_exist_for_sure: open this_doesnt_exist_for_sure: no such file or directory",
-			expPublicClusterAddress: "",
-		},
-		{
-			name: "using flag value to provoke error in SplitHostPort",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          "abc::123",
-			expErr:                  true,
-			expErrStr:               "Error splitting public cluster adddress host/port: address abc::123: too many colons in address",
-			expPublicClusterAddress: "",
-		},
-		{
-			name: "bad ip template",
-			inputConfig: &config.Config{
-				SharedConfig: &configutil.SharedConfig{
-					Listeners: []*listenerutil.ListenerConfig{},
-				},
-				Controller: &config.Controller{},
-			},
-			inputFlagValue:          "{{ somethingthatdoesntexist }}",
-			expErr:                  true,
-			expErrStr:               "Error parsing IP template on controller public cluster addr: unable to parse address template \"{{ somethingthatdoesntexist }}\": unable to parse template \"{{ somethingthatdoesntexist }}\": template: sockaddr.Parse:1: function \"somethingthatdoesntexist\" not defined",
-			expPublicClusterAddress: "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.stateFn != nil {
-				tt.stateFn(t)
-			}
-			s := Server{}
-			err := s.SetupControllerPublicClusterAddress(tt.inputConfig, tt.inputFlagValue)
-			if tt.expErr {
-				require.EqualError(t, err, tt.expErrStr)
-				return
-			}
-
-			require.NoError(t, err)
-			require.NotNil(t, tt.inputConfig.Controller)
-			require.Equal(t, tt.expPublicClusterAddress, tt.inputConfig.Controller.PublicClusterAddr)
-		})
-	}
-}
-
 func TestSetupWorkerPublicAddress(t *testing.T) {
 	t.Parallel()
 	tests := []struct {
